# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AverageRatingResponse:
    properties:
      average_rating:
        type: number
      reviewed_id:
        type: integer
      total_reviews:
        type: integer
    type: object
...
  dto.CreateReviewRequest:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: number
      reviewed_id:
        description: ID User yang akan diberikan review
        type: integer
    required:
    - comment
    - rating
    - reviewed_id
    type: object
...
  dto.ReviewResponse:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        type: number
      reviewed_id:
        type: integer
      reviewer_id:
        type: integer
    type: object
...
  dto.UpdateReviewRequest:
    properties:
      comment:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: number
    required:
    - comment
    - rating
    type: object
...
  utils.ErrorResponseSwagger:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk manajemen jobseeker
  title: Jobseeker Management API
  version: "1.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Create a new review. Only authenticated reviewer can create a review.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Create a new review
      tags:
      - reviews
  /reviews/{review_id}:
    delete:
      consumes:
      - application/json
      description: Delete a review. Only the reviewer who submitted the review
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review deleted successfully
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: 'Unauthorized: You can only delete your own reviews'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Delete Review
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update an existing review. Only the reviewer who submitted the
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: integer
      - description: Review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Review updated successfully
          schema:
            $ref: '#/definitions/dto.ReviewResponse'
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "401":
          description: 'Unauthorized: You can only update your own reviews'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Update Review
      tags:
      - reviews
  /reviews/average/{user_id}:
    get:
      consumes:
      - application/json
      description: Retrieve the average rating of a user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Average rating retrieved successfully
          schema:
            $ref: '#/definitions/dto.AverageRatingResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Get Average Rating
      tags:
      - reviews
  /reviews/me:
    get:
      consumes:
      - application/json
      description: Retrieve all reviews submitted by the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Reviews retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dto.ReviewResponse'
            type: array
        "401":
          description: 'Unauthorized: No user ID found in token'
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponseSwagger'
      security:
      - BearerAuth: []
      summary: Get My Reviews
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Masukkan token dalam format "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"